Fixed issue with my port file!!!!
So happy that I'm actually going to keep using that one
I think the one on the repo actually has better functionality so Ill keep it for now
using portb at the moment, it's a matter of pride now! lol

+++ portb.h
@@ -17,7 +17,7 @@
 		Port8Bit(uint16_t portnumber);
 		~Port8Bit();
 		virtual void Write(uint8_t data);
-		virtual uint8_t Read(uint8_t portnumber);
+		virtual uint8_t Read(uint16_t portnumber);
 	};
 
 	class Port8BitSlow : public Port8Bit
